+++ lib/nnq/cli/surveyor_respondent.rb
@@ -29,8 +29,8 @@ module NNQ
 
 
       def survey_and_collect(msg)
-        return if msg.empty?
-        body = config.format == :marshal ? Marshal.dump(msg.first) : msg.first
+        return if msg.nil? || msg.empty?
+        body = config.format == :marshal ? Marshal.dump(msg) : msg
         @sock.send_survey(body)
         transient_ready!
         collect_replies
@@ -42,7 +42,7 @@ module NNQ
           body = @sock.receive
           break if body.nil?
           reply = config.format == :marshal ? Marshal.load(body) : body
-          output(eval_recv_expr([reply]))
+          output(eval_recv_expr(reply))
         rescue NNQ::TimedOut
           break
         end
@@ -81,7 +81,7 @@ module NNQ
           reply = eval_recv_expr(msg)
           unless reply.equal?(SENT)
             output(reply)
-            send_reply(reply || [""])
+            send_reply(reply || "")
           end
         elsif config.echo
           output(msg)
@@ -99,8 +99,8 @@ module NNQ
 
 
       def send_reply(msg)
-        return if msg.empty?
-        body = config.format == :marshal ? Marshal.dump(msg.first) : msg.first
+        return if msg.nil?
+        body = config.format == :marshal ? Marshal.dump(msg) : msg
         @sock.send_reply(body)
         transient_ready!
       end